+++ src/Components/Card/CardHeader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Halo\UI\Components\Card;
+namespace Prism\UI\Components\Card;
 
 use Illuminate\View\Component;
 
@@ -10,6 +10,6 @@ class CardHeader extends Component
 {
     public function render()
     {
-        return view('components.halo.card.card-header');
+        return view('components.prism.card.card-header');
     }
 }