+++ src/Components/Modal/Modal.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Halo\UI\Components\Modal;
+namespace Prism\UI\Components\Modal;
 
 use Illuminate\View\Component;
 
@@ -17,6 +17,6 @@ class Modal extends Component
 
     public function render()
     {
-        return view('components.halo.modal.modal');
+        return view('components.prism.modal.modal');
     }
 }