+++ src/Components/Modal/ModalBody.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Halo\UI\Components\Modal;
+namespace Prism\UI\Components\Modal;
 
 use Illuminate\View\Component;
 
@@ -11,6 +11,6 @@ class ModalBody extends Component
 
     public function render()
     {
-        return view('components.halo.modal.modal-body');
+        return view('components.prism.modal.modal-body');
     }
 }